requirements.txt
fastapi
uvicorn
pandas 
requests 
pgvector
faiss-cpu
chromadb
pytest 
google-adk 
psycopg2-binary 